[RHELC-936] Port is_loaded_kernel_latest to Action framework (#778)
* [RHELC-936] Port is_loaded_kernel_latest to Action framework
