+++ convert2rhel/actions/__init__.py
@@ -49,29 +49,10 @@ BAD_KERNEL_RELEASE_SUBSTRINGS = ("uek", "rt", "linode")
 RPM_GPG_KEY_PATH = os.path.join(utils.DATA_DIR, "gpg-keys", "RPM-GPG-KEY-redhat-release")
 # The SSL certificate of the https://cdn.redhat.com/ server
 SSL_CERT_PATH = os.path.join(utils.DATA_DIR, "redhat-uep.pem")
-CDN_URL = "https://cdn.redhat.com/content/public/convert2rhel/$releasever/$basearch/os/"
-CONVERT2RHEL_REPO_CONTENT = """\
-[convert2rhel]
-name=Convert2RHEL Repository
-baseurl=%s
-gpgcheck=1
-enabled=1
-sslcacert=%s
-gpgkey=file://%s""" % (
-    CDN_URL,
-    SSL_CERT_PATH,
-    RPM_GPG_KEY_PATH,
-)
 
-PKG_NEVR = r"\b(\S+)-(?:([0-9]+):)?(\S+)-(\S+)\b"
 
 LINK_KMODS_RH_POLICY = "https://access.redhat.com/third-party-software-support"
 LINK_PREVENT_KMODS_FROM_LOADING = "https://access.redhat.com/solutions/41278"
-# The kernel version stays the same throughout a RHEL major version
-COMPATIBLE_KERNELS_VERS = {
-    7: "3.10.0",
-    8: "4.18.0",
-}
 
 #: Status code of an Action.
 #:
@@ -310,41 +291,10 @@ def run_actions():
     #
 
 
-def perform_system_checks():
-    """Early checks after system facts should be added here."""
-
-    is_loaded_kernel_latest()
-
-
 def perform_pre_ponr_checks():
     """Late checks before ponr should be added here."""
     ensure_compatibility_of_kmods()
     validate_package_manager_transaction()
-
-
-def check_tainted_kmods():
-    """Stop the conversion when a loaded tainted kernel module is detected.
-
-    Tainted kmods ends with (...) in /proc/modules, for example:
-        multipath 20480 0 - Live 0x0000000000000000
-        linear 20480 0 - Live 0x0000000000000000
-        system76_io 16384 0 - Live 0x0000000000000000 (OE)  <<<<<< Tainted
-        system76_acpi 16384 0 - Live 0x0000000000000000 (OE) <<<<<< Tainted
-    """
-    logger.task("Prepare: Check if loaded kernel modules are not tainted")
-    unsigned_modules, _ = run_subprocess(["grep", "(", "/proc/modules"])
-    module_names = "\n  ".join([mod.split(" ")[0] for mod in unsigned_modules.splitlines()])
-    if unsigned_modules:
-        logger.critical(
-            "Tainted kernel modules detected:\n  {0}\n"
-            "Third-party components are not supported per our "
-            "software support policy:\n {1}\n"
-            "Prevent the modules from loading by following {2}"
-            " and run convert2rhel again to continue with the conversion.".format(
-                module_names, LINK_KMODS_RH_POLICY, LINK_PREVENT_KMODS_FROM_LOADING
-            )
-        )
-    logger.info("No tainted kernel module is loaded.")
 
 
 def ensure_compatibility_of_kmods():
@@ -542,195 +492,3 @@ def get_unsupported_kmods(host_kmods, rhel_supported_kmods):
         for kmod in unsupported_kmods_subpaths
     ]
     return unsupported_kmods_full_paths
-
-
-def _bad_kernel_version(kernel_release):
-    """Return True if the booted kernel version does not correspond to the kernel version available in RHEL."""
-    kernel_version = kernel_release.split("-")[0]
-    try:
-        incompatible_version = COMPATIBLE_KERNELS_VERS[system_info.version.major] != kernel_version
-        if incompatible_version:
-            logger.warning(
-                "Booted kernel version '%s' does not correspond to the version "
-                "'%s' available in RHEL %d"
-                % (
-                    kernel_version,
-                    COMPATIBLE_KERNELS_VERS[system_info.version.major],
-                    system_info.version.major,
-                )
-            )
-        else:
-            logger.debug(
-                "Booted kernel version '%s' corresponds to the version available in RHEL %d"
-                % (kernel_version, system_info.version.major)
-            )
-        return incompatible_version
-    except KeyError:
-        logger.debug("Unexpected OS major version. Expected: %r" % COMPATIBLE_KERNELS_VERS.keys())
-        return True
-
-
-def _bad_kernel_package_signature(kernel_release):
-    """Return True if the booted kernel is not signed by the original OS vendor, i.e. it's a custom kernel."""
-    vmlinuz_path = "/boot/vmlinuz-%s" % kernel_release
-
-    kernel_pkg, return_code = run_subprocess(
-        ["rpm", "-qf", "--qf", "%{VERSION}&%{RELEASE}&%{ARCH}&%{NAME}", vmlinuz_path], print_output=False
-    )
-
-    os_vendor = system_info.name.split()[0]
-    if return_code == 1:
-        logger.warning(
-            "The booted kernel %s is not owned by any installed package."
-            " It needs to be owned by a package signed by %s." % (vmlinuz_path, os_vendor)
-        )
-
-        return True
-
-    version, release, arch, name = tuple(kernel_pkg.split("&"))
-    logger.debug("Booted kernel package name: {0}".format(name))
-
-    kernel_pkg_obj = get_installed_pkg_objects(name, version, release, arch)[0]
-    kernel_pkg_gpg_fingerprint = get_pkg_fingerprint(kernel_pkg_obj)
-    bad_signature = system_info.cfg_content["gpg_fingerprints"] != kernel_pkg_gpg_fingerprint
-
-    # e.g. Oracle Linux Server -> Oracle or
-    #      Oracle Linux Server -> CentOS Linux
-    if bad_signature:
-        logger.warning("Custom kernel detected. The booted kernel needs to be signed by %s." % os_vendor)
-        return True
-
-    logger.debug("The booted kernel is signed by %s." % os_vendor)
-    return False
-
-
-def _bad_kernel_substring(kernel_release):
-    """Return True if the booted kernel release contains one of the strings that identify it as non-standard kernel."""
-    bad_substring = any(bad_substring in kernel_release for bad_substring in BAD_KERNEL_RELEASE_SUBSTRINGS)
-    if bad_substring:
-        logger.debug(
-            "The booted kernel '{0}' contains one of the disallowed "
-            "substrings: {1}".format(kernel_release, BAD_KERNEL_RELEASE_SUBSTRINGS)
-        )
-        return True
-    return False
-
-
-def is_loaded_kernel_latest():
-    """Check if the loaded kernel is behind or of the same version as in yum repos."""
-    logger.task("Prepare: Check if the loaded kernel version is the most recent")
-
-    if system_info.id == "oracle" and system_info.corresponds_to_rhel_eus_release():
-        logger.info(
-            "Skipping the check because there are no publicly available %s %d.%d repositories available."
-            % (system_info.name, system_info.version.major, system_info.version.minor)
-        )
-        return
-
-    cmd = [
-        "repoquery",
-        "--setopt=exclude=",
-        "--quiet",
-        "--qf",
-        "C2R\\t%{BUILDTIME}\\t%{VERSION}-%{RELEASE}\\t%{REPOID}",
-    ]
-
-    reposdir = get_hardcoded_repofiles_dir()
-    if reposdir and not system_info.has_internet_access:
-        logger.warning("Skipping the check as no internet connection has been detected.")
-        return
-
-    # If the reposdir variable is not empty, meaning that it detected the
-    # hardcoded repofiles, we should use that
-    # instead of the system repositories located under /etc/yum.repos.d
-    if reposdir:
-        cmd.append("--setopt=reposdir=%s" % reposdir)
-
-    # For Oracle/CentOS Linux 8 the `kernel` is just a meta package, instead,
-    # we check for `kernel-core`. But 7 releases, the correct way to check is
-    # using `kernel`.
-    package_to_check = "kernel-core" if system_info.version.major >= 8 else "kernel"
-
-    # Append the package name as the last item on the list
-    cmd.append(package_to_check)
-
-    # Look up for available kernel (or kernel-core) packages versions available
-    # in different repositories using the `repoquery` command.  If convert2rhel
-    # detects that it is running on a EUS system, then repoquery will use the
-    # hardcoded repofiles available under `/usr/share/convert2rhel/repos`,
-    # meaning that the tool will fetch only the latest kernels available for
-    # that EUS version, and not the most updated version from other newer
-    # versions.
-    repoquery_output, return_code = run_subprocess(cmd, print_output=False)
-    if return_code != 0:
-        logger.debug("Got the following output: %s", repoquery_output)
-        logger.warning(
-            "Couldn't fetch the list of the most recent kernels available in "
-            "the repositories. Skipping the loaded kernel check."
-        )
-        return
-
-    packages = []
-    # We are expecting a repoquery output which is similar to this:
-    #   C2R     1671212820      3.10.0-1160.81.1.el7    updates
-    # We need the `C2R` identifier to be present on the line so we can know for
-    # sure that the line we are working with is a line that contains
-    # relevant repoquery information to our check, otherwise, we just log the
-    # information as debug and do nothing with it.
-    for line in repoquery_output.split("\n"):
-        if line.strip() and "C2R" in line:
-            _, build_time, latest_kernel, repoid = tuple(str(line).split("\t"))
-            packages.append((build_time, latest_kernel, repoid))
-        else:
-            # Mainly for debugging purposes to see what is happening if we got
-            # anything else that does not have the C2R identifier at the start
-            # of the line.
-            logger.debug("Got a line without the C2R identifier: %s" % line)
-
-    # If we don't have any packages, then something went wrong, we need to
-    # decide wether to bail out or output a warning (only if the user used the
-    # special environment variable for it.
-    if not packages:
-        unsupported_skip = os.environ.get("CONVERT2RHEL_UNSUPPORTED_SKIP_KERNEL_CURRENCY_CHECK", None)
-        if not unsupported_skip:
-            logger.critical(
-                "Could not find any %s from repositories to compare against the loaded kernel.\n"
-                "Please, check if you have any vendor repositories enabled to proceed with the conversion.\n"
-                "If you wish to ignore this message, set the environment variable "
-                "'CONVERT2RHEL_UNSUPPORTED_SKIP_KERNEL_CURRENCY_CHECK' to 1." % package_to_check
-            )
-
-        logger.warning(
-            "Detected 'CONVERT2RHEL_UNSUPPORTED_SKIP_KERNEL_CURRENCY_CHECK' environment variable, we will skip "
-            "the %s comparison.\n"
-            "Beware, this could leave your system in a broken state. " % package_to_check
-        )
-        return
-
-    packages.sort(key=lambda x: x[0], reverse=True)
-    _, latest_kernel, repoid = packages[0]
-
-    uname_output, _ = run_subprocess(["uname", "-r"], print_output=False)
-    loaded_kernel = uname_output.rsplit(".", 1)[0]
-    # append the package name to loaded_kernel and latest_kernel so they can be properly processed by
-    # compare_package_versions()
-    latest_kernel_pkg = "%s-%s" % (package_to_check, latest_kernel)
-    loaded_kernel_pkg = "%s-%s" % (package_to_check, loaded_kernel)
-    match = compare_package_versions(latest_kernel_pkg, loaded_kernel_pkg)
-
-    if match != 0:
-        repos_message = (
-            "in the enabled system repositories"
-            if not reposdir
-            else "in repositories defined in the %s folder" % reposdir
-        )
-        logger.critical(
-            "The version of the loaded kernel is different from the latest version %s.\n"
-            " Latest kernel version available in %s: %s\n"
-            " Loaded kernel version: %s\n\n"
-            "To proceed with the conversion, update the kernel version by executing the following step:\n\n"
-            "1. yum install %s-%s -y\n"
-            "2. reboot" % (repos_message, repoid, latest_kernel, loaded_kernel, package_to_check, latest_kernel)
-        )
-
-    logger.info("The currently loaded kernel is at the latest version.")
