+++ convert2rhel/unit_tests/actions/rhel_compatible_kernel_test.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright(C) 2018 Red Hat, Inc.
+# Copyright(C) 2023 Red Hat, Inc.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
